Data.List.Utils

Documentation

chunks :: Int -> [a] -> [[a]]

Break up a list into chunks of size n, or < n if there are < n elements (remaining) in the list

 λ> chunks 1 [1..10]
 [[1],[2],[3],[4],[5],[6],[7],[8],[9],[10],[]]

 λ> chunks 2 [1..10]
 [[1,2],[3,4],[5,6],[7,8],[9,10],[]]

 λ> chunks 3 [1..10]
 [[1,2,3],[4,5,6],[7,8,9],[10]]

alternate :: Num a => [a] -> [a]

Negate all numbers at odd indices

For example:

 λ> alternate [0..10]
 [0,-1,2,-3,4,-5,6,-7,8,-9,10]

autoAlternate :: (Num a, Ord a) => [a] -> [a]

alternate if signums of the first two elements differ

For example:

It adds alternation to [0..10] because (signum 0 == 0) /= (signum 1 == 1):

 λ> autoAlternate [0..10]
 [0,-1,2,-3,4,-5,6,-7,8,-9,10]

And removes alternation from an already alternating list:

 λ> autoAlternate $ alternate [0..10]
 [0,1,2,3,4,5,6,7,8,9,10]

diff :: Num a => [a] -> [a]

First (forward) differences:

 λ> diff [1..10]
 [1,1,1,1,1,1,1,1,1]

 λ> diff [1,3..10]
 [2,2,2,2]

autoDiff :: (Ord a, Num a) => [a] -> [a]

Automatically alternate the first differences

  autoAlternate . diff

takeExactly :: Int -> [a] -> Maybe [a]

Take exactly the given number of elements or return Nothing

findCycle :: Eq a => [a] -> Maybe [a]

Find a subsequence ys of xs such that:

 and $ zipWith (==) xs (cycle ys)

isCycle :: Eq a => [a] -> [a] -> Bool

Is the second argument equal to:

 take n $ cycle xs

for some n?

 λ> all (n -> isCycle [1,2] (take n $ cycle [1, 2])) [2..2^12]
 True
 (0.53 secs, 1,884,383,408 bytes)

 λ> any (n -> isCycle [1,2] (take n $ 2 : cycle [1, 2])) [2..2^12]
 False
 (0.01 secs, 4,943,968 bytes)

boolMaybe :: (a -> Bool) -> a -> Maybe a

Just if the predicate returns True

For example, use with mapMaybe to filter a list:

 λ> mapMaybe (boolMaybe odd) [1..10]
 [1,3,5,7,9]

Should be equivalent to:

 \p -> liftM2 (>>) ((`unless` Nothing) . p) return