Data.Poly

Documentation

newtype Poly a

Univariate polynomials

Constructors

Poly

Fields

unPoly :: Free PolyF a

var :: Poly a

Poly variable

toPoly :: (Eq a, Fractional a) => [a] -> Maybe (Poly a)

Convert a list to a Poly if possible

toPolys :: (Eq a, Fractional a) => [a] -> [Poly a]

Convert a list to a list of Polys

toPolyN :: (Eq a, Fractional a) => Int -> [a] -> Maybe (Poly a)

Convert to Poly if possible using the given number of elements

interpolate :: Fractional a => [a] -> Poly a

Interpolate a list of Fractional values into a Poly

Currently buggy:

 λ> runPoly (interpolate [1,2,3]) $ [0..3]
 [1.0,1.0,1.0,1.0]
 (0.01 secs, 485,672 bytes)

simplify :: Fractional a => Poly a -> Poly a

Simplify combinations of constants and x + x

freeLevels :: (Functor f, Eq1 f, Eq a) => (Free f a -> Free f a) -> Free f a -> Free f a

Apply a transformation on a Free Monad on all levels, repeating until equivalence using fixEq

fixEq :: Eq a => (a -> a) -> a -> a

Apply the function until it reaches a fixed point

simplifyAll :: (Fractional a, Eq a) => Poly a -> Poly a

Examples:

 λ> simplifyAll (1 + 2 + 3 :: Poly Double)
 Poly {unPoly = Pure 6.0}

 λ> simplifyAll (1 + 2 + 3 + var :: Poly Double)
 Poly {unPoly = Free (:+ (Pure 6.0) (Free Var))}

 λ> simplifyAll (1 + 2 + 3 * (var + var) :: Poly Double)
 Poly {unPoly = Free (:+ (Pure 3.0) (Free (:* (Pure 3.0) (Free (:* (Pure 2.0) (Free Var))))))}

wrap2 :: MonadFree m f => (a -> b -> m (f c)) -> a -> b -> f c

wrap after applying a binary function

runPoly :: Fractional a => Poly a -> a -> a

Evaluate a univariate polynomial on the given input